Rename webpack config files
